+++ epyk/core/html/HtmlList.py
@@ -2,6 +2,8 @@
 # -*- coding: utf-8 -*-
 
 """
+Description:
+-----------
 List of all the different templates configurations available for displaying bespoke lists.
 This list can be extended and it is easy to test a new configuration by different defining the HTML template in the
 common list object.