+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Integrate Moment library to the ase Javascript features.
 - Improve documentation.
+- Add packages white / black lists shortcuts.
 
 ### Changed